voice-guard/requirements.txt
# Web Framework
fastapi>=0.111
uvicorn[standard]>=0.30

# Database
SQLAlchemy>=2.0
pymysql>=1.1

# Google Cloud Services
google-cloud-speech==2.21.0
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-genai

# AWS Services
boto3>=1.34

# WebSocket & File Upload
websockets==12.0
python-multipart==0.0.6

# Environment & Data Validation
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.2